correccion de funciones para horarios

+++ controllers/horarioController.js
@@ -2,17 +2,45 @@ const Horario = require('../models/Horario');
 const Pelicula = require('../models/Pelicula');
 const Sala = require('../models/Sala');
 const PDFDocument = require('pdfkit');
+const { Op } = require('sequelize');const { validationResult } = require('express-validator');
 
 // Crear un nuevo horario
 exports.createHorario = async (req, res) => {
-    const { idPelicula, idSala, horaProgramada, fechaDeEmision } = req.body;
+    const { horaProgramada, fechaDeEmision, turno } = req.body;
+
+    // Validación de datos
+    if (!horaProgramada) {
+        return res.status(400).json({ message: 'La hora programada es requerida' });
+    }
+    if (!fechaDeEmision) {
+        return res.status(400).json({ message: 'La fecha de emisión es requerida' });
+    }
+    if (!turno) {
+        return res.status(400).json({ message: 'El turno es requerido' });
+    }
+
+    // Validar el formato de horaProgramada
+    if (!/^([01]\d|2[0-3]):([0-5]\d):([0-5]\d)$/.test(horaProgramada)) {
+        return res.status(400).json({ message: 'Formato de hora no válido' });
+    }
+
+    // Validar el formato de fechaDeEmision
+    const formattedDate = new Date(fechaDeEmision);
+    if (isNaN(formattedDate.getTime())) {
+        return res.status(400).json({ message: 'Formato de fecha no válido' });
+    }
+
+    // Validar el turno
+    const validTurnos = ['mañana', 'tarde', 'noche'];
+    if (!validTurnos.includes(turno)) {
+        return res.status(400).json({ message: 'Turno no válido' });
+    }
 
     try {
         const horario = await Horario.create({
-            idPelicula,
-            idSala,
             horaProgramada,
-            fechaDeEmision
+            fechaDeEmision,
+            turno
         });
 
         res.status(201).json(horario);
@@ -20,6 +48,7 @@ exports.createHorario = async (req, res) => {
         res.status(500).json({ error: error.message });
     }
 };
+
 
 // Obtener todos los horarios
 exports.getAllHorarios = async (req, res) => {
@@ -34,14 +63,48 @@ exports.getAllHorarios = async (req, res) => {
 // Obtener horarios por criterios de búsqueda
 exports.getHorarios = async (req, res) => {
     try {
-        const { idHorario, idPelicula, idSala, fechaDeEmision } = req.query;
+        const { idHorario, fechaDeEmision, horaProgramada, turno } = req.query;
         const searchCriteria = {};
-        if (idHorario) searchCriteria.idHorario = idHorario;
-        if (idPelicula) searchCriteria.idPelicula = idPelicula;
-        if (idSala) searchCriteria.idSala = idSala;
-        if (fechaDeEmision) searchCriteria.fechaDeEmision = fechaDeEmision;
 
-        const horarios = await Horario.findAll({ where: searchCriteria });
+        if (idHorario) searchCriteria.idHorario = idHorario;
+
+        if (fechaDeEmision) {
+            // Convertir fecha a formato YYYY-MM-DD
+            const formattedDate = new Date(fechaDeEmision);
+            if (isNaN(formattedDate.getTime())) {
+                return res.status(400).json({ message: 'Formato de fecha no válido' });
+            }
+            searchCriteria.fechaDeEmision = {
+                [Op.eq]: formattedDate.toISOString().split('T')[0] // Formato YYYY-MM-DD
+            };
+        }
+
+        if (horaProgramada) {
+            // Validar y ajustar formato de horaProgramada
+            const time = horaProgramada;
+            if (!/^([01]\d|2[0-3]):([0-5]\d):([0-5]\d)$/.test(time)) {
+                return res.status(400).json({ message: 'Formato de hora no válido' });
+            }
+            searchCriteria.horaProgramada = {
+                [Op.eq]: time
+            };
+        }
+
+        if (turno) {
+            if (!['mañana', 'tarde', 'noche'].includes(turno)) {
+                return res.status(400).json({ message: 'Turno no válido' });
+            }
+            searchCriteria.turno = turno;
+        }
+
+        console.log('Search Criteria:', searchCriteria); // Depura los criterios de búsqueda
+
+        // Ejecutar la consulta
+        const horarios = await Horario.findAll({
+            where: searchCriteria
+        });
+
+        console.log('Horarios Encontrados:', horarios); // Depura los horarios encontrados
 
         if (horarios.length > 0) {
             res.json(horarios);
@@ -55,45 +118,80 @@ exports.getHorarios = async (req, res) => {
 
 // Actualizar horarios por múltiples criterios
 exports.updateHorarios = async (req, res) => {
-    try {
-        const { idHorario, idPelicula, idSala, horaProgramada, fechaDeEmision } = req.body;
+    const { idHorario, horaProgramada, fechaDeEmision, turno } = req.body;
 
-        if (!idHorario) {
-            return res.status(400).json({ message: 'ID del horario es requerido para la actualización' });
+    // Validación de datos
+    if (!idHorario) {
+        return res.status(400).json({ message: 'ID del horario es requerido para la actualización' });
+    }
+
+    // Validar el formato de horaProgramada, si se proporciona
+    if (horaProgramada && !/^([01]\d|2[0-3]):([0-5]\d):([0-5]\d)$/.test(horaProgramada)) {
+        return res.status(400).json({ message: 'Formato de hora no válido' });
+    }
+
+    // Validar el formato de fechaDeEmision, si se proporciona
+    if (fechaDeEmision) {
+        const formattedDate = new Date(fechaDeEmision);
+        if (isNaN(formattedDate.getTime())) {
+            return res.status(400).json({ message: 'Formato de fecha no válido' });
+        }
+    }
+
+    // Validar el turno, si se proporciona
+    if (turno && !['mañana', 'tarde', 'noche'].includes(turno)) {
+        return res.status(400).json({ message: 'Turno no válido' });
+    }
+
+    try {
+        // Encuentra el horario a actualizar
+        const horario = await Horario.findByPk(idHorario);
+        if (!horario) {
+            return res.status(404).json({ message: 'Horario no encontrado' });
         }
 
-        const updateFields = {};
-        if (idPelicula) updateFields.idPelicula = idPelicula;
-        if (idSala) updateFields.idSala = idSala;
-        if (horaProgramada) updateFields.horaProgramada = horaProgramada;
-        if (fechaDeEmision) updateFields.fechaDeEmision = fechaDeEmision;
+        // Actualizar el horario con los campos proporcionados
+        const updated = await Horario.update(
+            {
+                horaProgramada: horaProgramada || horario.horaProgramada,
+                fechaDeEmision: fechaDeEmision || horario.fechaDeEmision,
+                turno: turno || horario.turno
+            },
+            { where: { idHorario } }
+        );
 
-        const [updated] = await Horario.update(updateFields, { where: { idHorario } });
-
-        if (updated) {
+        if (updated[0] === 1) {
             res.json({ message: 'Horario actualizado exitosamente' });
         } else {
-            res.status(404).json({ message: 'Horario no encontrado' });
+            res.status(400).json({ message: 'No se pudo actualizar el horario' });
         }
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
 };
 
-// Eliminar un horario
 exports.deleteHorario = async (req, res) => {
     try {
         const { idHorario } = req.params;
+        console.log('ID del horario recibido:', idHorario);
 
         if (!idHorario) {
             return res.status(400).json({ message: 'ID del horario es requerido para la eliminación' });
         }
 
+        // Verificar si el horario existe
         const horario = await Horario.findByPk(idHorario);
         if (!horario) {
             return res.status(404).json({ message: 'Horario no encontrado' });
         }
 
+        // Verificar si hay registros en la tabla peliculas asociados a este horario
+        const peliculas = await Pelicula.findAll({ where: { idHorario } });
+        if (peliculas.length > 0) {
+            return res.status(400).json({ message: 'No se puede eliminar el horario porque hay registros asociados en la tabla peliculas' });
+        }
+
+        // Eliminar el horario
         await Horario.destroy({ where: { idHorario } });
 
         res.json({ message: 'Horario eliminado exitosamente' });
@@ -105,12 +203,7 @@ exports.deleteHorario = async (req, res) => {
 // Generar reporte en PDF de los horarios
 exports.getHorariosPDF = async (req, res) => {
     try {
-        const horarios = await Horario.findAll({
-            include: [
-                { model: Pelicula, attributes: ['nombrePelicula'] },
-                { model: Sala, attributes: ['nombreSala'] }
-            ]
-        });
+        const horarios = await Horario.findAll();
 
         const doc = new PDFDocument();
 
@@ -123,10 +216,9 @@ exports.getHorariosPDF = async (req, res) => {
 
         horarios.forEach(horario => {
             doc.fontSize(12).text(`ID: ${horario.idHorario}`);
-            doc.fontSize(12).text(`Película: ${horario.Pelicula.nombrePelicula}`);
-            doc.fontSize(12).text(`Sala: ${horario.Sala.nombreSala}`);
             doc.fontSize(12).text(`Hora Programada: ${horario.horaProgramada}`);
-            doc.fontSize(12).text(`Fecha de Emision: ${horario.fechaDeEmision}`);
+            doc.fontSize(12).text(`Turno: ${horario.turno}`);
+            doc.fontSize(12).text(`Fecha de Emisión: ${horario.fechaDeEmision}`);
             doc.moveDown();
         });
